+++ apps/docs/icons/nextjs.tsx
@@ -1,0 +1,59 @@
+export default () => (
+  <svg
+    width="32"
+    height="32"
+    viewBox="0 0 32 32"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <mask
+      id="mask0_24_13476"
+      style={{ maskType: "alpha" }}
+      maskUnits="userSpaceOnUse"
+      x="0"
+      y="0"
+      width="32"
+      height="32"
+    >
+      <circle cx="16" cy="16" r="16" fill="black" />
+    </mask>
+    <g mask="url(#mask0_24_13476)">
+      <circle cx="16" cy="16" r="16" fill="black" />
+      <path
+        d="M26.5792 28.0036L12.2919 9.59998H9.59998V22.3946H11.7535V12.3348L24.8887 29.3058C25.4814 28.9091 26.046 28.4738 26.5792 28.0036Z"
+        fill="url(#paint0_linear_24_13476)"
+      />
+      <rect
+        x="20.4445"
+        y="9.59998"
+        width="2.13333"
+        height="12.8"
+        fill="url(#paint1_linear_24_13476)"
+      />
+    </g>
+    <defs>
+      <linearGradient
+        id="paint0_linear_24_13476"
+        x1="19.3778"
+        y1="20.7111"
+        x2="25.6889"
+        y2="28.5333"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+      <linearGradient
+        id="paint1_linear_24_13476"
+        x1="21.5111"
+        y1="9.59998"
+        x2="21.4754"
+        y2="19"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stop-color="white" />
+        <stop offset="1" stop-color="white" stop-opacity="0" />
+      </linearGradient>
+    </defs>
+  </svg>
+);