+++ packages/react/src/components/alert-dialog.tsx
@@ -21,6 +21,7 @@ import {
   createContext,
   forwardRef,
   ForwardRefExoticComponent,
+  FunctionComponent,
   useContext,
 } from "react";
 import { clx } from "../utils";
@@ -38,27 +39,31 @@ const AlertDialogContext = createContext<AlertDialogContextProps>({
   variant: "default",
 });
 
-interface AlertDialogProps
+export interface AlertDialogProps
   extends ComponentProps<typeof Dialog>,
     AlertDialogContextProps {}
 
-const AlertDialog: ForwardRefExoticComponent<AlertDialogProps> = forwardRef(
-  ({ variant, ...props }) => {
-    return (
-      <AlertDialogContext.Provider value={{ variant }}>
-        <Dialog {...props} />
-      </AlertDialogContext.Provider>
-    );
-  },
-);
-
-interface AlertDialogActionProps
+const AlertDialog: FunctionComponent<AlertDialogProps> = ({
+  variant,
+  ...props
+}) => {
+  return (
+    <AlertDialogContext.Provider value={{ variant }}>
+      <Dialog {...props} />
+    </AlertDialogContext.Provider>
+  );
+};
+export interface AlertDialogActionProps
   extends Omit<ComponentProps<typeof DialogFooter>, "border"> {}
 
 const AlertDialogAction: ForwardRefExoticComponent<AlertDialogActionProps> =
-  forwardRef(({ children, className, ...props }) => {
+  forwardRef(({ children, className, ...props }, ref) => {
     return (
-      <DialogFooter className={clx("px-0 pb-0", className)} {...props}>
+      <DialogFooter
+        ref={ref}
+        className={clx("px-0 pb-0", className)}
+        {...props}
+      >
         {children}
       </DialogFooter>
     );
@@ -103,6 +108,14 @@ const AlertDialogContent: ForwardRefExoticComponent<
     </DialogBody>
   );
 });
+
+AlertDialog.displayName = "AlertDialog";
+AlertDialogContent.displayName = "AlertDialogContent";
+AlertDialogDescription.displayName = "AlertDialogDescription";
+AlertDialogAction.displayName = "AlertDialogAction";
+AlertDialogClose.displayName = "AlertDialogClose";
+AlertDialogTitle.displayName = "AlertDialogTitle";
+AlertDialogTrigger.displayName = "AlertDialogTrigger";
 
 export {
   AlertDialog,