+++ packages/react/src/components/data-table.tsx
@@ -583,5 +583,6 @@ const DataTable = <TData extends Record<string, any>>(
     </DataTableProvider>
   );
 };
+DataTable.displayName = "DataTable";
 
 export { DataTable, Cell, type ColumnDef };