+++ packages/react/src/components/date-field.tsx
@@ -95,5 +95,6 @@ const DateField: FunctionComponent<DateFieldProps> = ({
     </div>
   );
 };
+DateField.displayName = "DateField";
 
 export { DateField };