+++ packages/react/src/components/date-picker.tsx
@@ -68,8 +68,8 @@ const DatePicker: FC<DatePickerProps> = ({
             {date ? formatDate(date) : placeholder}
           </Button>
         </DialogTrigger>
-        <DialogBody hideClose>
-          <DialogContent className="w-fit p-0">
+        <DialogBody hideClose className="min-w-fit w-fit">
+          <DialogContent className="p-0">
             <Calendar
               className="w-[304px]"
               disabled={disabled}