+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @govtechmy/myds-react
 
+## 0.0.9
+
+### Patch Changes
+
+- f50b423: Fix: Make asChild non-default
+- 2f6d05d: Fix peer dependency version to include react 19
+
+## 0.0.8
+
+### Patch Changes
+
+- a61fba9: Add `hideClose` prop to DialogBody
+- 26ed81e: Masthead css update (container)
+- c45e030: remove DropdownTrigger `asChild` set as always true
+
 ## 0.0.7
 
 ### Patch Changes